requirements.txt
pandas 
numpy
ipykernel
kagglehub
pyyaml